APIs/pressassociation.io/2.0/python-aiohttp-client/tests/test_feature_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web



pytestmark = pytest.mark.asyncio

async def test_get_feature(client):
    """Test case for get_feature

    Feature Detail
    """
    headers = { 
        'Accept': 'application/json',
        'apikey': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v2/feature/{feature_id}'.format(feature_id='feature_id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_feature_types(client):
    """Test case for list_feature_types

    Feature Type Collection
    """
    headers = { 
        'Accept': 'application/json',
        'apikey': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v2/feature-type',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_features(client):
    """Test case for list_features

    Feature Collection
    """
    params = [('type', 'netflix-monthly'),
                    ('date', '2018-09-15'),
                    ('start', '2018-09-15'),
                    ('end', '2018-10-15')]
    headers = { 
        'Accept': 'application/json',
        'apikey': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v2/feature',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

